Split round robin logic into separate modules
