+++ gomoku/scripts/__init__.py
@@ -1,1 +1,16 @@
-"""学習や対戦などのスクリプトを格納するサブパッケージ"""
+"""学習や対戦用の各種スクリプトをまとめたサブパッケージ"""
+
+from .tournament_core import play_match, round_robin_tournament
+from .tournament_plot import (
+    print_tournament_results,
+    plot_winrate_heatmap,
+    plot_ranking_bar,
+)
+
+__all__ = [
+    "play_match",
+    "round_robin_tournament",
+    "print_tournament_results",
+    "plot_winrate_heatmap",
+    "plot_ranking_bar",
+]
